Move convertDefaultDisplays to dashboard utils
