+++ gsa/src/web/components/sortable/utils.js
@@ -20,10 +20,6 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-import uuid from 'uuid/v4';
-
-import {createDisplay, createRow} from 'gmp/commands/dashboards';
-
 export const removeItem = (rows, itemId) => rows.map(row => ({
   ...row,
   items: row.items.filter(item => item.id !== itemId),
@@ -33,16 +29,5 @@ export const updateRow = (row, data) => ({
   ...row,
   ...data,
 });
-
-export const convertDefaultDisplays = (defaultDisplays = [],
-   uuidFunc = uuid) => {
-  return {
-    rows: defaultDisplays.map(row => createRow(
-      row.map(displayId => createDisplay(displayId, undefined, uuidFunc)),
-      undefined,
-      uuidFunc
-    )),
-  };
-};
 
 // vim: set ts=2 sw=2 tw=80:
